refactor(bus,cli): split bus.rs and cli_smoke.rs to satisfy 300-line cap
CI's file-size-guard step refused commit 4e7cabc with two files over
the 300-line cap (CONSTITUTION § 13). Both grew during the merge of
main into wave-0b. Splits introduced:

- crates/convergio-bus/src/bus.rs (was 322): extracted
  publish_system + poll_system (ADR-0024, system.* topics with
  plan_id IS NULL) into a new sibling file
  crates/convergio-bus/src/bus_system.rs. The two methods are
  defined as `impl Bus` in the new file and access bus.rs internals
  via three pub(crate) hooks: a Bus::pool() accessor, MessageRow,
  and next_seq. Result: bus.rs is 258 lines, bus_system.rs is 88.
  No public API changed.

- crates/convergio-cli/tests/cli_smoke.rs (was 336): the two Wave 0b
  setup-extras tests (`setup_agent_claude_emits_skill_and_settings`
  and `setup_agent_copilot_does_not_emit_claude_extras`) moved to a
  new sibling file cli_smoke_setup.rs (49 lines). The
  `status_help_lists_agents_flag` test moved to the existing
  cli_smoke_status.rs alongside the other status help tests.
  cli_smoke.rs is now 295 lines.

Local verification:
  cargo fmt --all -- --check     clean
  cargo clippy --workspace -D    clean
  cargo test --workspace         all green (cli + bus + server)
  file-size guard                no file above 300

Tracks plan 2564b354-… (Wave 0b — Claude Code adapter), task
ccc3093d-… (w1.10 — Open Wave 0b PR + ensure CI green).
