+++ crates/convergio-bus/src/bus.rs
@@ -1,13 +1,13 @@
 //! Bus — write/read API for Layer 2.
 
 use crate::error::{BusError, Result};
-use crate::model::{Message, NewMessage, NewSystemMessage, TopicSummary};
+use crate::model::{Message, NewMessage, TopicSummary};
 use chrono::{DateTime, Utc};
 use convergio_db::Pool;
 use uuid::Uuid;
 
 /// Topic prefix that marks the system-scoped family (ADR-0024).
-const SYSTEM_TOPIC_PREFIX: &str = "system.";
+pub(crate) const SYSTEM_TOPIC_PREFIX: &str = "system.";
 
 /// Read/write access to the message bus.
 #[derive(Clone)]
@@ -20,6 +20,11 @@ impl Bus {
     /// [`crate::init`] at least once.
     pub fn new(pool: Pool) -> Self {
         Self { pool }
+    }
+
+    /// Pool accessor for sibling impl blocks (e.g. [`crate::bus_system`]).
+    pub(crate) fn pool(&self) -> &Pool {
+        &self.pool
     }
 
     /// Append a plan-scoped message to the bus. The topic MUST NOT
@@ -67,50 +72,6 @@ impl Bus {
         })
     }
 
-    /// Append a system-scoped message (ADR-0024). The topic MUST start
-    /// with `system.`; rejects otherwise. Stored with `plan_id IS NULL`.
-    pub async fn publish_system(&self, msg: NewSystemMessage) -> Result<Message> {
-        if !msg.topic.starts_with(SYSTEM_TOPIC_PREFIX) {
-            return Err(BusError::InvalidTopicScope(format!(
-                "topic '{}' is not system-scoped; use publish",
-                msg.topic
-            )));
-        }
-        let payload = serde_json::to_string(&msg.payload)?;
-        let id = Uuid::new_v4().to_string();
-        let now = Utc::now();
-        let now_str = now.to_rfc3339();
-        let mut tx = self.pool.inner().begin().await?;
-        let next_seq = next_seq(&mut tx).await?;
-
-        sqlx::query(
-            "INSERT INTO agent_messages \
-             (id, seq, plan_id, topic, sender, payload, consumed_at, consumed_by, created_at) \
-             VALUES (?, ?, NULL, ?, ?, ?, NULL, NULL, ?)",
-        )
-        .bind(&id)
-        .bind(next_seq)
-        .bind(&msg.topic)
-        .bind(&msg.sender)
-        .bind(&payload)
-        .bind(&now_str)
-        .execute(&mut *tx)
-        .await?;
-        tx.commit().await?;
-
-        Ok(Message {
-            id,
-            seq: next_seq,
-            plan_id: None,
-            topic: msg.topic,
-            sender: msg.sender,
-            payload: msg.payload,
-            consumed_at: None,
-            consumed_by: None,
-            created_at: now,
-        })
-    }
-
     /// Poll unconsumed messages for `(plan_id, topic)` since `cursor`
     /// (exclusive). Returns up to `limit` rows in `seq` order.
     ///
@@ -140,31 +101,6 @@ impl Bus {
              ORDER BY seq ASC LIMIT ?",
         )
         .bind(plan_id)
-        .bind(topic)
-        .bind(cursor)
-        .bind(limit)
-        .fetch_all(self.pool.inner())
-        .await?;
-        rows.into_iter().map(TryInto::try_into).collect()
-    }
-
-    /// Poll unconsumed system-scoped messages for `topic` since
-    /// `cursor` (exclusive). The topic MUST start with `system.`;
-    /// rejects otherwise. See ADR-0024.
-    pub async fn poll_system(&self, topic: &str, cursor: i64, limit: i64) -> Result<Vec<Message>> {
-        if !topic.starts_with(SYSTEM_TOPIC_PREFIX) {
-            return Err(BusError::InvalidTopicScope(format!(
-                "topic '{topic}' is not system-scoped; use poll"
-            )));
-        }
-        let rows = sqlx::query_as::<_, MessageRow>(
-            "SELECT id, seq, plan_id, topic, sender, payload, consumed_at, \
-                    consumed_by, created_at \
-             FROM agent_messages \
-             WHERE plan_id IS NULL AND topic = ? AND seq > ? \
-                   AND consumed_at IS NULL \
-             ORDER BY seq ASC LIMIT ?",
-        )
         .bind(topic)
         .bind(cursor)
         .bind(limit)
@@ -269,7 +205,7 @@ impl Bus {
     }
 }
 
-async fn next_seq(tx: &mut sqlx::Transaction<'_, sqlx::Sqlite>) -> Result<i64> {
+pub(crate) async fn next_seq(tx: &mut sqlx::Transaction<'_, sqlx::Sqlite>) -> Result<i64> {
     let row: (i64,) = sqlx::query_as(
         "UPDATE agent_message_sequence SET next_seq = next_seq + 1 \
          WHERE id = 1 RETURNING next_seq - 1",
@@ -280,7 +216,7 @@ async fn next_seq(tx: &mut sqlx::Transaction<'_, sqlx::Sqlite>) -> Result<i64> {
 }
 
 #[derive(sqlx::FromRow)]
-struct MessageRow {
+pub(crate) struct MessageRow {
     id: String,
     seq: i64,
     plan_id: Option<String>,
